Fix compilation on Windows (clang-cl)

+++ test_runner_lib/include/test_runner/TestRunner.h
@@ -45,7 +45,7 @@ namespace test_runner {
 
 
 using random_bytes_engine =
-    std::independent_bits_engine<std::default_random_engine, CHAR_BIT, uint8_t>;
+    std::independent_bits_engine<std::default_random_engine, CHAR_BIT, uint16_t>;
 
 
 class MemoryHandler {
